refactor: rename and refactor getAndTranslateData
The file was previously name translateData. Decided
to move getTimeEntries into this file for readability.
Updated all imports for this change.
